Move TrackList to its own component
